Updated Promotions BE delete logic and FE  (#97)
* Check if promotions tied to transactions before relevant delete actions

* Edited promo fe to match other pages

* added additional checks to ensure purchase transaction tied to credit bundle

* added PromotionStatusEnum in frontend promotion type

+++ apps/backend/src/services/PromotionService.ts
@@ -1,5 +1,6 @@
 import { Promotion, Prisma } from '@prisma/client';
 import PromotionDao from '../dao/PromotionDao';
+import TransactionService from './TransactionService';
 
 class PromotionService {
   public async createPromotion(
@@ -31,8 +32,13 @@ class PromotionService {
     return PromotionDao.updatePromotion(promotionId, data);
   }
 
-  public async deletePromotion(promotionId: string): Promise<Promotion> {
-    return PromotionDao.deletePromotion(promotionId);
+  public async deletePromotion(promotionId: string): Promise<Promotion | null> {
+    const transactions = await TransactionService.getTransactionsByPromotionId(
+      promotionId,
+    );
+    if (transactions.length === 0) {
+      return PromotionDao.deletePromotion(promotionId);
+    } else return PromotionDao.softDeletePromotion(promotionId);
   }
 }
